backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1"
description = "interactive knowledge graph builder"
package-mode = false

[tool.poetry.dependencies]
python = "3.13.7"
arcadepy = "^1.8.0"
pydantic-ai = "^1.0.10"